requirements-dev.txt
# Development dependencies
pytest==7.4.3
requests-mock==1.11.0
ipython==8.17.2
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Advanced scraping (optional)
selenium==4.15.2

# Include production requirements
-r requirements.txt